+++ apps/api/src/app/order/interfaces/order-with-account.type.ts
@@ -1,3 +1,5 @@
-import { Account, Order } from '@prisma/client';
+import { Account, Order, Platform } from '@prisma/client';
 
-export type OrderWithAccount = Order & { Account?: Account };
+type AccountWithPlatform = Account & { Platform?: Platform };
+
+export type OrderWithAccount = Order & { Account?: AccountWithPlatform };